Rename Red Sky API to just API
